Add play loyout module
