+++ packages/client/src/modules/play/PlayLayout/index.tsx
@@ -1,0 +1,3 @@
+import { PlayLayout } from "./PlayLayout";
+
+export { PlayLayout };
